docs(README): document required token permissions
Under the default restrictive access for GitHub tokens, this action is
not able to operate, and the API will throw a "Resource not accessible
by integration" error.
The user has to either expand the access of the default token on all
workflows, or explicitly grant write access to repo contents to the
actioin step.

https://docs.github.com/en/actions/security-guides/automatic-token-authentication#permissions-for-the-github_token